Fix runtime lib clippy lints.
Resolve the strict clippy diagnostics in the runtime library:
- allow upper-case acronyms for CPU/FPU/MMX/XMM,
- replace Flags::to_string with Display,
- introduce EXEData::Block type alias,
- clean up needless returns, needless borrows, lazy evaluations,
- use is_multiple_of, remove identity shifts, remove unnecessary casts,
- use slice copy_from_slice in pslldq/psrldq,
- use iter_mut/enumerate in palignr_xmm,
- avoid negated float comparisons with partial_cmp/matches!.

`cargo clippy -p runtime --lib -- -D warnings` and `cargo test -p runtime`
now pass.

+++ runtime/src/registers.rs
@@ -200,7 +200,7 @@ impl Regs {
     }
 
     pub fn get_dx_ax(&self) -> u32 {
-        ((self.edx as u32) << 16) | self.eax as u32
+        (self.edx << 16) | self.eax
     }
 
     pub fn set_dx_ax(&mut self, val: u32) {
